Move parsing code to its own file
