+++ src/rs/parser.rs
@@ -1,0 +1,493 @@
+/* SPDX-FileCopyrightText: © 2025 Decompollaborate */
+/* SPDX-License-Identifier: MIT */
+
+use std::{collections::HashSet, path::Path};
+
+use regex::*;
+
+use crate::{mapfile::MapFile, section, segment, symbol, utils};
+
+lazy_static! {
+    static ref BANNED_SYMBOL_NAMES: HashSet<&'static str> = {
+        let mut symbol_names = HashSet::new();
+        symbol_names.insert("gcc2_compiled.");
+        symbol_names
+    };
+}
+
+impl MapFile {
+    pub fn new() -> Self {
+        Self {
+            segments_list: Vec::new(),
+
+            #[cfg(feature = "python_bindings")]
+            debugging: false,
+        }
+    }
+
+    /// Creates a new `MapFile` object and fills it with the contents from the
+    /// file pointed by the `map_path` argument.
+    ///
+    /// The format of the map will be guessed based on its contents.
+    ///
+    /// Currently supported map formats:
+    /// - GNU ld
+    /// - clang ld.lld
+    pub fn new_from_map_file(map_path: &Path) -> Self {
+        let mut m = Self::new();
+        m.read_map_file(map_path);
+        m
+    }
+
+    /**
+    Opens the mapfile pointed by the `map_path` argument and parses it.
+
+    The format of the map will be guessed based on its contents.
+
+    Currently supported map formats:
+    - GNU ld
+    - clang ld.lld
+     */
+    pub fn read_map_file(&mut self, map_path: &Path) {
+        let map_contents = utils::read_file_contents(map_path);
+
+        self.parse_map_contents(&map_contents);
+    }
+
+    /**
+    Parses the contents of the map.
+
+    The `map_contents` argument must contain the contents of a mapfile.
+
+    The format of the map will be guessed based on its contents.
+
+    Currently supported mapfile formats:
+    - GNU ld
+    - clang ld.lld
+    */
+    pub fn parse_map_contents(&mut self, map_contents: &str) {
+        let regex_lld_header =
+            Regex::new(r"\s+VMA\s+LMA\s+Size\s+Align\s+Out\s+In\s+Symbol").unwrap();
+
+        if regex_lld_header.is_match(map_contents) {
+            self.parse_map_contents_lld(map_contents);
+        } else {
+            // GNU is the fallback
+            self.parse_map_contents_gnu(map_contents);
+        }
+    }
+
+    /**
+    Parses the contents of a GNU ld map.
+
+    The `map_contents` argument must contain the contents of a GNU ld mapfile.
+     */
+    pub fn parse_map_contents_gnu(&mut self, map_contents: &str) {
+        // TODO: maybe move somewhere else?
+        let regex_section_alone_entry = Regex::new(r"^\s+(?P<section>[^*][^\s]+)\s*$").unwrap();
+        let regex_section_data_entry = Regex::new(r"^\s+(?P<section>([^*][^\s]+)?)\s+(?P<vram>0x[^\s]+)\s+(?P<size>0x[^\s]+)\s+(?P<name>[^\s]+)$").unwrap();
+        let regex_function_entry =
+            Regex::new(r"^\s+(?P<vram>0x[^\s]+)\s+(?P<name>[^\s]+)$").unwrap();
+        // regex_function_entry = re.compile(r"^\s+(?P<vram>0x[^\s]+)\s+(?P<name>[^\s]+)((\s*=\s*(?P<expression>.+))?)$")
+        let regex_label = Regex::new(r"(?P<name>\.?L[0-9A-F]{8})$").unwrap();
+        let regex_fill =
+            Regex::new(r"^\s+(?P<fill>\*[^\s\*]+\*)\s+(?P<vram>0x[^\s]+)\s+(?P<size>0x[^\s]+)\s+(?P<fillValue>[0-9a-zA-Z]*)$")
+                .unwrap();
+        let regex_segment_entry = Regex::new(r"(?P<name>([^\s]+)?)\s+(?P<vram>0x[^\s]+)\s+(?P<size>0x[^\s]+)\s+(?P<loadaddress>(load address)?)\s+(?P<vrom>0x[^\s]+)$").unwrap();
+
+        let map_data = MapFile::preprocess_map_data_gnu(map_contents);
+
+        let mut temp_segment_list = vec![segment::Segment::new_placeholder()];
+
+        let mut in_section = false;
+
+        let mut prev_line = "";
+        for line in map_data.split('\n') {
+            if in_section {
+                if !line.starts_with("        ") {
+                    in_section = false;
+                } else if !regex_label.is_match(line) {
+                    // Filter out jump table's labels
+
+                    // Find symbols
+                    if let Some(entry_match) = regex_function_entry.captures(line) {
+                        // println!("{entry_match:?}");
+                        let sym_name = &entry_match["name"];
+
+                        if !BANNED_SYMBOL_NAMES.contains(&sym_name) {
+                            let sym_vram = utils::parse_hex(&entry_match["vram"]);
+
+                            let current_segment = temp_segment_list.last_mut().unwrap();
+                            let current_section = current_segment.sections_list.last_mut().unwrap();
+
+                            current_section
+                                .symbols
+                                .push(symbol::Symbol::new_default(sym_name.into(), sym_vram));
+                        }
+                    }
+                }
+            }
+
+            if !in_section {
+                if let Some(section_entry_match) = regex_section_data_entry.captures(line) {
+                    let filepath = std::path::PathBuf::from(&section_entry_match["name"]);
+                    let vram = utils::parse_hex(&section_entry_match["vram"]);
+                    let size = utils::parse_hex(&section_entry_match["size"]);
+                    let section_type = &section_entry_match["section"];
+
+                    if size > 0 {
+                        // TODO: de-duplicate the following code:
+
+                        if !section_type.is_empty() {
+                            in_section = true;
+                            let current_segment = temp_segment_list.last_mut().unwrap();
+
+                            current_segment
+                                .sections_list
+                                .push(section::Section::new_default(
+                                    filepath,
+                                    vram,
+                                    size,
+                                    section_type,
+                                ));
+                        } else if let Some(section_alone_match) =
+                            regex_section_alone_entry.captures(prev_line)
+                        {
+                            // Some sections may be too large, making the entry be splitted between two lines, making the section name be in one line and the rest of the info in the next one
+
+                            let section_type = &section_alone_match["section"];
+
+                            in_section = true;
+                            let current_segment = temp_segment_list.last_mut().unwrap();
+
+                            current_segment
+                                .sections_list
+                                .push(section::Section::new_default(
+                                    filepath,
+                                    vram,
+                                    size,
+                                    section_type,
+                                ));
+                        }
+                    }
+                } else if let Some(segment_entry_match) = regex_segment_entry.captures(line) {
+                    let mut name = &segment_entry_match["name"];
+                    let vram = utils::parse_hex(&segment_entry_match["vram"]);
+                    let size = utils::parse_hex(&segment_entry_match["size"]);
+                    let vrom = utils::parse_hex(&segment_entry_match["vrom"]);
+
+                    if name.is_empty() {
+                        // If the segment name is too long then this line gets break in two lines
+                        name = prev_line;
+                    }
+
+                    temp_segment_list.push(segment::Segment::new_default(
+                        name.into(),
+                        vram,
+                        size,
+                        vrom,
+                    ));
+                } else if let Some(fill_match) = regex_fill.captures(line) {
+                    // Make a dummy file to handle *fill*
+                    let mut filepath = std::path::PathBuf::new();
+                    let mut vram = 0;
+                    let size = utils::parse_hex(&fill_match["size"]);
+                    let mut section_type = "".to_owned();
+
+                    let current_segment = temp_segment_list.last_mut().unwrap();
+
+                    if !current_segment.sections_list.is_empty() {
+                        let prev_section = current_segment.sections_list.last().unwrap();
+                        let mut name = prev_section.filepath.file_name().unwrap().to_owned();
+
+                        name.push("__fill__");
+                        filepath = prev_section.filepath.with_file_name(name);
+                        vram = prev_section.vram + prev_section.size;
+                        section_type.clone_from(&prev_section.section_type);
+                    }
+
+                    current_segment
+                        .sections_list
+                        .push(section::Section::new_fill(
+                            filepath,
+                            vram,
+                            size,
+                            &section_type,
+                        ));
+                }
+            }
+
+            prev_line = line;
+        }
+
+        self.segments_list = Self::post_process_segments_gnu(temp_segment_list);
+    }
+
+    fn post_process_segments_gnu(
+        temp_segment_list: Vec<segment::Segment>,
+    ) -> Vec<segment::Segment> {
+        let mut segments_list = Vec::with_capacity(temp_segment_list.len());
+
+        for (i, segment) in temp_segment_list.into_iter().enumerate() {
+            if i == 0 && segment.is_placeholder() {
+                // skip the dummy segment if it has no size, sections or symbols
+                continue;
+            }
+
+            let mut new_segment = segment.clone_no_sectionlist();
+
+            let mut vrom_offset = segment.vrom;
+            for mut section in segment.sections_list.into_iter() {
+                let mut acummulated_size = 0;
+                let symbols_count = section.symbols.len();
+                let is_noload_section = section.is_noload_section();
+
+                if section.is_placeholder() {
+                    // drop placeholders
+                    continue;
+                }
+
+                if section.vrom.is_some() {
+                    vrom_offset = section.vrom.unwrap();
+                }
+
+                if !is_noload_section {
+                    section.vrom = Some(vrom_offset);
+                }
+
+                if symbols_count > 0 {
+                    let mut sym_vrom = vrom_offset;
+
+                    // The first symbol of the section on the mapfile may not be the actual first
+                    // symbol if it is marked `static`, be a jumptable, etc, producing a mismatch
+                    // on the vrom address of each symbol of this section.
+                    // A way to adjust this difference is by increasing the start of the vrom
+                    // by the difference in vram address between the first symbol and the vram
+                    // of the section.
+                    if let Some(first_sym) = section.symbols.first() {
+                        sym_vrom += first_sym.vram - section.vram;
+
+                        // Aditionally, if the first symbol is missing then calculation of the size
+                        // for the last symbol would be wrong, since we subtract the accumulated
+                        // size of each symbol from the section's total size to calculate it.
+                        // We need to adjust the total size by this difference too.
+                        acummulated_size += first_sym.vram - section.vram;
+                    }
+
+                    // Calculate size of each symbol
+                    for index in 0..symbols_count - 1 {
+                        let next_sym_vram = section.symbols[index + 1].vram;
+                        let sym = &mut section.symbols[index];
+                        let sym_size = next_sym_vram - sym.vram;
+                        acummulated_size += sym_size;
+
+                        sym.size = sym_size;
+
+                        if !is_noload_section {
+                            // Only set vrom of non bss variables
+                            sym.vrom = Some(sym_vrom);
+                            sym_vrom += sym_size;
+                        }
+                    }
+
+                    // Calculate size of last symbol of the section
+                    let sym = &mut section.symbols[symbols_count - 1];
+                    let sym_size = section.size - acummulated_size;
+                    sym.size = sym_size;
+                    if !is_noload_section {
+                        sym.vrom = Some(sym_vrom);
+                        //sym_vrom += sym_size;
+                    }
+                }
+
+                if !is_noload_section {
+                    vrom_offset += section.size;
+                }
+
+                new_segment.sections_list.push(section);
+            }
+
+            segments_list.push(new_segment);
+        }
+
+        segments_list.shrink_to_fit();
+        segments_list
+    }
+
+    /**
+    Parses the contents of a clang ld.lld map.
+
+    The `map_contents` argument must contain the contents of a clang ld.lld mapfile.
+     */
+    pub fn parse_map_contents_lld(&mut self, map_contents: &str) {
+        let map_data = map_contents;
+
+        // Every line starts with this information, so instead of duplicating it we put them on one single regex
+        let regex_row_entry = Regex::new(r"^\s*(?P<vram>[0-9a-fA-F]+)\s+(?P<vrom>[0-9a-fA-F]+)\s+(?P<size>[0-9a-fA-F]+)\s+(?P<align>[0-9a-fA-F]+) ").unwrap();
+
+        let regex_segment_entry = Regex::new(r"^(?P<name>[^\s]+)$").unwrap();
+        let regex_fill = Regex::new(r"^\s+(?P<expr>\.\s*\+=\s*.+)$").unwrap();
+        let regex_section_data_entry =
+            Regex::new(r"^\s+(?P<name>[^\s]+):\((?P<section>[^\s()]+)\)$$").unwrap();
+        let regex_label = Regex::new(r"^\s+(?P<name>\.?L[0-9A-F]{8})$").unwrap();
+        let regex_symbol_entry = Regex::new(r"^\s+(?P<name>[^\s]+)$").unwrap();
+
+        let mut temp_segment_list = vec![segment::Segment::new_placeholder()];
+
+        for line in map_data.split('\n') {
+            if let Some(row_entry_match) = regex_row_entry.captures(line) {
+                let vram = utils::parse_hex(&row_entry_match["vram"]);
+                let vrom = utils::parse_hex(&row_entry_match["vrom"]);
+                let size = utils::parse_hex(&row_entry_match["size"]);
+                let align = utils::parse_hex(&row_entry_match["align"]);
+
+                let subline = &line[row_entry_match.get(0).unwrap().len()..];
+
+                if let Some(segment_entry_match) = regex_segment_entry.captures(subline) {
+                    let name = &segment_entry_match["name"];
+
+                    let mut new_segment =
+                        segment::Segment::new_default(name.into(), vram, size, vrom);
+                    new_segment.align = Some(align);
+
+                    temp_segment_list.push(new_segment);
+                } else if regex_fill.is_match(subline) {
+                    // Make a dummy section to handle pads (. += XX)
+
+                    let mut filepath = std::path::PathBuf::new();
+                    let mut section_type = "".to_owned();
+
+                    let current_segment = temp_segment_list.last_mut().unwrap();
+
+                    if !current_segment.sections_list.is_empty() {
+                        let prev_section = current_segment.sections_list.last().unwrap();
+                        let mut name = prev_section.filepath.file_name().unwrap().to_owned();
+
+                        name.push("__fill__");
+                        filepath = prev_section.filepath.with_file_name(name);
+                        section_type.clone_from(&prev_section.section_type);
+                    }
+
+                    let mut new_section =
+                        section::Section::new_fill(filepath, vram, size, &section_type);
+                    if !utils::is_noload_section(&section_type) {
+                        new_section.vrom = Some(vrom);
+                    }
+                    current_segment.sections_list.push(new_section);
+                } else if let Some(section_entry_match) = regex_section_data_entry.captures(subline)
+                {
+                    let filepath = std::path::PathBuf::from(&section_entry_match["name"]);
+                    let section_type = &section_entry_match["section"];
+
+                    if size > 0 {
+                        let current_segment = temp_segment_list.last_mut().unwrap();
+
+                        let mut new_section =
+                            section::Section::new_default(filepath, vram, size, section_type);
+                        if !utils::is_noload_section(section_type) {
+                            new_section.vrom = Some(vrom);
+                        }
+                        new_section.align = Some(align);
+
+                        current_segment.sections_list.push(new_section);
+                    }
+                } else if regex_label.is_match(subline) {
+                    // pass
+                } else if let Some(symbol_entry_match) = regex_symbol_entry.captures(subline) {
+                    let name = &symbol_entry_match["name"];
+
+                    if !BANNED_SYMBOL_NAMES.contains(&name) {
+                        let current_segment = temp_segment_list.last_mut().unwrap();
+                        let current_section = current_segment.sections_list.last_mut().unwrap();
+
+                        let mut new_symbol = symbol::Symbol::new_default(name.into(), vram);
+                        if size > 0 {
+                            new_symbol.size = size;
+                        }
+                        if !current_section.is_noload_section() {
+                            new_symbol.vrom = Some(vrom)
+                        }
+                        new_symbol.align = Some(align);
+
+                        current_section.symbols.push(new_symbol);
+                    }
+                }
+            }
+        }
+
+        self.segments_list = Self::post_process_segments_lld(temp_segment_list);
+    }
+
+    fn post_process_segments_lld(
+        temp_segment_list: Vec<segment::Segment>,
+    ) -> Vec<segment::Segment> {
+        let mut segments_list = Vec::with_capacity(temp_segment_list.len());
+
+        for (i, segment) in temp_segment_list.into_iter().enumerate() {
+            if i == 0 && segment.is_placeholder() {
+                // skip the dummy segment if it has no size, sections or symbols
+                continue;
+            }
+
+            let mut new_segment = segment.clone_no_sectionlist();
+
+            for mut section in segment.sections_list.into_iter() {
+                if section.is_placeholder() {
+                    // drop placeholders
+                    continue;
+                }
+
+                let mut acummulated_size = 0;
+                let symbols_count = section.symbols.len();
+
+                if symbols_count > 0 {
+                    // Calculate the size of symbols that the map section did not report.
+                    // usually asm symbols and not C ones
+
+                    for index in 0..symbols_count - 1 {
+                        let next_sym_vram = section.symbols[index + 1].vram;
+                        let sym = &mut section.symbols[index];
+
+                        let sym_size = next_sym_vram - sym.vram;
+                        acummulated_size += sym_size;
+
+                        if sym.size == 0 {
+                            sym.size = sym_size;
+                        }
+                    }
+
+                    // Calculate size of last symbol of the section
+                    let sym = &mut section.symbols[symbols_count - 1];
+                    if sym.size == 0 {
+                        let sym_size = section.size - acummulated_size;
+                        sym.size = sym_size;
+                    }
+                }
+
+                new_segment.sections_list.push(section);
+            }
+
+            segments_list.push(new_segment);
+        }
+
+        segments_list.shrink_to_fit();
+        segments_list
+    }
+}
+
+impl MapFile {
+    // TODO: figure out if this is doing unnecessary copies or something
+    fn preprocess_map_data_gnu(map_data: &str) -> String {
+        // Skip the stuff we don't care about
+        // Looking for this string will only work on English machines (or C locales)
+        // but it doesn't matter much, because if this string is not found then the
+        // parsing should still work, but just a bit slower because of the extra crap
+        if let Some(aux_var) = map_data.find("\nLinker script and memory map") {
+            if let Some(start_index) = map_data[aux_var + 1..].find('\n') {
+                return map_data[aux_var + 1 + start_index + 1..].to_string();
+            }
+        }
+
+        map_data.to_string()
+    }
+}
